Replacing int with NSInteger for 64-bit compatibility

+++ UIMonthYearPicker/UIMonthYearPicker.m
@@ -103,8 +103,8 @@ const NSInteger numberOfComponents = 2;
 
 - (void) setDate:(NSDate *)aDate animated:(BOOL)animated{
     
-    int dateYear;
-    int dateMonth;
+    NSInteger dateYear;
+    NSInteger dateMonth;
     
     NSDateComponents *components = [[NSCalendar currentCalendar] components:NSMonthCalendarUnit fromDate:aDate];
     dateMonth = [components month];
@@ -115,7 +115,7 @@ const NSInteger numberOfComponents = 2;
         inComponent: MONTH
            animated: NO];
     
-    int rowIndex = dateYear - minYear;
+    NSInteger rowIndex = dateYear - minYear;
     if (rowIndex < 0) {
         rowIndex = 0;
     }
